Reduces `Temperature` integer precision
Reduces precision from (-2^27, 2^27 - 1) to (-2^11, 2^11 - 1).
Or (-134217728, 134217727) to (-2048, 2047).

Decimal precision remains the same.

+++ src/ds18b20.rs
@@ -114,7 +114,7 @@ impl Ds18b20 {
         }
 
         let value = i16::from_le_bytes([buf[0], buf[1]]);
-        Ok(Temperature::from_bits(i32::from(value)))
+        Ok(Temperature::from_bits(value))
     }
 
     /// Asynchronously Measures the temperature
